Setup intergation test project.
