+++ .github/workflows/dotnet.yml
@@ -16,8 +16,6 @@ jobs:
     - name: Start Database with Docker Compose
       run: docker compose -f docker-compose-test.yml up -d qa-platform-db
 
-    - uses: actions/checkout@v4
-
     - name: Start API Container
       run: docker compose -f docker-compose-test.yml up -d qa-platform-api
