+++ QAPlatformAPI.IntegrationTests/IntegrationTestBase.cs
