--- QAPlatformAPI.IntegrationTests/appsettings.Testing.json
@@ -1,19 +1,0 @@
-{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft.AspNetCore": "Warning"
-    }
-  },
-  "ConnectionStrings": {
-    "PostgreSQLConnection": "Server=qa-platform-db;Port=5433;Database=qa-platform-db;Username=devuser;Password=devpassword"
-  },
-  "CORS": {
-    "AllowedOrigins": [ "*" ]
-  },
-  "JwtSettings": {
-    "Issuer": "qa-platform.ltu.se",
-    "Audience": "qa-platform.ltu.se",
-    "Expires": 60000
-  }
-}
