+++ QAPlatformAPI/Program.cs
@@ -1,4 +1,3 @@
-using Microsoft.Extensions.Configuration;
 using QAPlatformAPI.Extensions;
 
 namespace QAPlatformAPI;
